+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New post-treater to set data to a constant value where under a threshold.
+- Updating `ForwardPower` or `ReflectedPower` automatically updates `SWR` and `ReflectionCoefficient`.
+  - Used `Observer` design pattern.
 
 ## [1.7.2] - 2025-06-03